requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.7
PyJWT==2.9.0
prometheus-client==0.20.0
alembic==1.13.2
phonenumbers==8.13.43
httpx==0.27.0
pytest==8.2.2
PyNaCl==1.5.0
